🐛 Fix claude image api format
Claude Image probably never worked. This commit fix claude vision.

# Why?
It never worked because the BasicMemoryAgent, when inserting images into the message list, it does so in OpenAI API format, which is different from how Claude does it.

I fixed it by adding a converter to convert OpenAI format of image message to Claude api format.